ISTO-128 Code review fixes
- Fix syndication feed sort NPE on missing contentItemVersion.
- Return unmodifiable copy from getAllPublishedVersions().
- Guard getTitleCleaned() against null and titles without a hyphen.
- Use ConcurrentHashMap for publishedBranches cache.
- Mock SyndicationService in security test to avoid MLDS calls.
- Fall back to default size when syndication length is malformed.
- Use long for copyWithProgress byte counter to avoid overflow on large downloads.

+++ src/main/java/org/snomed/snowstorm/core/util/StreamUtil.java
@@ -21,15 +21,15 @@ public class StreamUtil {
 		throw new IllegalStateException(String.format("Duplicate key %s", u.getConceptId()));
 	};
 
-	public static int copyWithProgress(InputStream inputStream, OutputStream outputStream, int totalStreamLength, String messageFormat) throws IOException {
+	public static long copyWithProgress(InputStream inputStream, OutputStream outputStream, int totalStreamLength, String messageFormat) throws IOException {
 		long basis = totalStreamLength <= 0 ? 1L : totalStreamLength;
 		return copyWithProgress(inputStream, outputStream, basis, messageFormat, null);
 	}
 
-	public static int copyWithProgress(InputStream inputStream, OutputStream outputStream, long totalStreamLength,
+	public static long copyWithProgress(InputStream inputStream, OutputStream outputStream, long totalStreamLength,
 			String messageFormat, IntConsumer onPercent) throws IOException {
 		long basis = totalStreamLength <= 0 ? 1L : totalStreamLength;
-		int byteCount = 0;
+		long byteCount = 0;
 		int bytesRead;
 		int percentageLogged = -1;
 		int lastCallbackPercent = -1;
